Rename Copilot markdown elements directory
