+++ spx-gui/src/components/copilot/CopilotRoot.vue
@@ -10,8 +10,8 @@ import { useIsRouteLoaded } from '@/utils/route-loading'
 import * as projectApis from '@/apis/project'
 import { useMessageEvents, useModalEvents } from '@/components/ui'
 import { Copilot, type ICopilotContextProvider, type SessionExported, type ToolDefinition } from './copilot'
-import * as pageLink from './custom-elements/PageLink'
-import * as highlightLink from './custom-elements/HighlightLink.vue'
+import * as pageLink from './markdown-elements/PageLink'
+import * as highlightLink from './markdown-elements/HighlightLink.vue'
 import { useSignedInStateQuery, type SignedInState } from '@/stores/user'
 import { userSessionStorageRef } from '@/utils/user-storage'
 import { provideCopilot } from './context'
