+++ spx-gui/src/components/copilot/markdown-elements/CodeBlock.vue
