+++ spx-gui/src/components/copilot/markdown-elements/HighlightLink.vue
