+++ spx-gui/src/components/copilot/markdown-elements/Thinking.ts
