+++ spx-gui/src/components/copilot/markdown-elements/common/BlockActionBtn.vue
