+++ spx-gui/src/components/copilot/markdown-elements/common/BlockFooter.vue
