+++ spx-gui/src/components/copilot/markdown-elements/common/BlockWrapper.vue
