+++ spx-gui/src/components/editor/copilot/CodeChange.vue
@@ -32,9 +32,9 @@ import { useMessageHandle, ActionException } from '@/utils/exception'
 import CodeView from '@/components/common/CodeView.vue'
 import { useEditorCtxRef } from '@/components/editor/EditorContextProvider.vue'
 import { CodeLink, getTextDocumentId, type Range, useCodeEditorRef } from '@/components/editor/spx-code-editor'
-import BlockWrapper from '@/components/copilot/custom-elements/common/BlockWrapper.vue'
-import BlockFooter from '@/components/copilot/custom-elements/common/BlockFooter.vue'
-import BlockActionBtn from '@/components/copilot/custom-elements/common/BlockActionBtn.vue'
+import BlockWrapper from '@/components/copilot/markdown-elements/common/BlockWrapper.vue'
+import BlockFooter from '@/components/copilot/markdown-elements/common/BlockFooter.vue'
+import BlockActionBtn from '@/components/copilot/markdown-elements/common/BlockActionBtn.vue'
 
 const props = defineProps<{
   /** Code file path, e.g., `NiuXiaoQi.spx` */
